+++ api/src/main/kotlin/com/few/api/domain/subscription/SubscribeWorkbookUseCase.kt
@@ -1,32 +1,32 @@
-package com.few.api.web.usecase.subscription
+package com.few.api.domain.subscription
 
 import com.few.api.repo.dao.subscription.SubscriptionDao
-import com.few.api.repo.dao.subscription.command.UpdateDeletedAtInWorkbookSubscriptionCommand
+import com.few.api.repo.dao.subscription.command.InsertWorkbookSubscriptionCommand
 import com.few.api.repo.dao.subscription.query.CountWorkbookSubscriptionQuery
-import com.few.api.web.usecase.subscription.`in`.UnsubscribeWorkbookUseCaseIn
+import com.few.api.domain.subscription.`in`.SubscribeWorkbookUseCaseIn
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
 
 @Component
-class UnsubscribeWorkbookUseCase(
+class SubscribeWorkbookUseCase(
     private val subscriptionDao: SubscriptionDao
 ) {
 
     @Transactional
-    fun execute(useCaseIn: UnsubscribeWorkbookUseCaseIn) {
+    fun execute(useCaseIn: SubscribeWorkbookUseCaseIn) {
         // TODO: request sending email
 
-        // 구독중이 아닌지 확인
+        // 이미 구독중인지 확인
         CountWorkbookSubscriptionQuery(memberId = useCaseIn.memberId, workbookId = useCaseIn.workbookId).let { query ->
             subscriptionDao.selectCountWorkbookSubscription(query).let { cnt ->
-                if (cnt == 0) {
+                if (cnt > 0) {
                     throw RuntimeException("Already subscribed")
                 }
             }
         }
 
-        subscriptionDao.updateDeletedAtInWorkbookSubscription(
-            UpdateDeletedAtInWorkbookSubscriptionCommand(memberId = useCaseIn.memberId, workbookId = useCaseIn.workbookId)
+        subscriptionDao.insertWorkbookSubscription(
+            InsertWorkbookSubscriptionCommand(memberId = useCaseIn.memberId, workbookId = useCaseIn.workbookId)
         )
     }
 }
