+++ api/src/main/kotlin/com/few/api/domain/subscription/in/SubscribeWorkbookUseCaseIn.kt
@@ -1,4 +1,4 @@
-package com.few.api.web.usecase.subscription.`in`
+package com.few.api.domain.subscription.`in`
 
 data class SubscribeWorkbookUseCaseIn(
     val workbookId: Long,
