+++ api/src/main/kotlin/com/few/api/web/controller/problem/ProblemController.kt
@@ -6,10 +6,10 @@ import com.few.api.web.controller.problem.response.ProblemContents
 import com.few.api.web.controller.problem.response.ReadProblemResponse
 import com.few.api.web.support.ApiResponse
 import com.few.api.web.support.ApiResponseGenerator
-import com.few.api.web.usecase.problem.CheckProblemUseCase
-import com.few.api.web.usecase.problem.ReadProblemUseCase
-import com.few.api.web.usecase.problem.`in`.CheckProblemUseCaseIn
-import com.few.api.web.usecase.problem.`in`.ReadProblemUseCaseIn
+import com.few.api.domain.problem.CheckProblemUseCase
+import com.few.api.domain.problem.ReadProblemUseCase
+import com.few.api.domain.problem.`in`.CheckProblemUseCaseIn
+import com.few.api.domain.problem.`in`.ReadProblemUseCaseIn
 import org.springframework.http.HttpStatus
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.*
