+++ api/src/main/kotlin/com/few/api/web/controller/subscription/WorkbookSubscriptionController.kt
@@ -4,10 +4,10 @@ import com.few.api.web.controller.subscription.request.SubscribeWorkbookRequest
 import com.few.api.web.controller.subscription.request.UnsubscribeWorkbookRequest
 import com.few.api.web.support.ApiResponse
 import com.few.api.web.support.ApiResponseGenerator
-import com.few.api.web.usecase.subscription.SubscribeWorkbookUseCase
-import com.few.api.web.usecase.subscription.UnsubscribeWorkbookUseCase
-import com.few.api.web.usecase.subscription.`in`.SubscribeWorkbookUseCaseIn
-import com.few.api.web.usecase.subscription.`in`.UnsubscribeWorkbookUseCaseIn
+import com.few.api.domain.subscription.SubscribeWorkbookUseCase
+import com.few.api.domain.subscription.UnsubscribeWorkbookUseCase
+import com.few.api.domain.subscription.`in`.SubscribeWorkbookUseCaseIn
+import com.few.api.domain.subscription.`in`.UnsubscribeWorkbookUseCaseIn
 import org.springframework.http.HttpStatus
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.*
