+++ api/src/test/kotlin/com/few/api/web/controller/problem/ProblemControllerTest.kt
@@ -9,13 +9,13 @@ import com.few.api.web.controller.ControllerTestSpec
 import com.few.api.web.controller.description.Description
 import com.few.api.web.controller.helper.*
 import com.few.api.web.controller.problem.request.CheckProblemBody
-import com.few.api.web.usecase.problem.CheckProblemUseCase
-import com.few.api.web.usecase.problem.ReadProblemUseCase
-import com.few.api.web.usecase.problem.`in`.CheckProblemUseCaseIn
-import com.few.api.web.usecase.problem.`in`.ReadProblemUseCaseIn
-import com.few.api.web.usecase.problem.out.CheckProblemUseCaseOut
-import com.few.api.web.usecase.problem.out.ReadProblemContentsUseCaseOutDetail
-import com.few.api.web.usecase.problem.out.ReadProblemUseCaseOut
+import com.few.api.domain.problem.CheckProblemUseCase
+import com.few.api.domain.problem.ReadProblemUseCase
+import com.few.api.domain.problem.`in`.CheckProblemUseCaseIn
+import com.few.api.domain.problem.`in`.ReadProblemUseCaseIn
+import com.few.api.domain.problem.out.CheckProblemUseCaseOut
+import com.few.api.domain.problem.out.ReadProblemContentsUseCaseOutDetail
+import com.few.api.domain.problem.out.ReadProblemUseCaseOut
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
